Handle default properties. Map improvements.
We no longer store properties in the db if they have the default values. Make the realmEditor and the game play code check this and read the defaults from the module data.

Make the realm editor show the object and character icons on the map, and the start location (using the player location icon).